+++ audit/asserts_test.go
@@ -116,12 +116,15 @@ func TestAssertAbout(t *testing.T) {
 func TestAssertRange(t *testing.T) {
 	successfulAssert := successfulAssertion(t)
 	failingAssert := failingAssertion(t)
+	now := time.Now()
 
 	successfulAssert.Range(byte(9), byte(1), byte(22), "byte in range")
 	successfulAssert.Range(9, 1, 22, "int in range")
 	successfulAssert.Range(9.0, 1.0, 22.0, "float64 in range")
 	successfulAssert.Range('f', 'a', 'z', "rune in range")
 	successfulAssert.Range("foo", "a", "zzzzz", "string in range")
+	successfulAssert.Range(now, now.Add(-time.Hour), now.Add(time.Hour), "time in range")
+	successfulAssert.Range(time.Minute, time.Second, time.Hour, "duration in range")
 	successfulAssert.Range([]int{1, 2, 3}, 1, 10, "slice length in range")
 	successfulAssert.Range([3]int{1, 2, 3}, 1, 10, "array length in range")
 	successfulAssert.Range(map[int]int{3: 1, 2: 2, 1: 3}, 1, 10, "map length in range")
@@ -130,6 +133,8 @@ func TestAssertRange(t *testing.T) {
 	failingAssert.Range(1.0, 10.0, 20.0, "float64 out of range")
 	failingAssert.Range('a', 'x', 'z', "rune out of range")
 	failingAssert.Range("aaa", "uuuuu", "zzzzz", "string out of range")
+	failingAssert.Range(now, now.Add(time.Minute), now.Add(time.Hour), "time out of range")
+	failingAssert.Range(time.Second, time.Minute, time.Hour, "duration in range")
 	failingAssert.Range([]int{1, 2, 3}, 5, 10, "slice length out of range")
 	failingAssert.Range([3]int{1, 2, 3}, 5, 10, "array length out of range")
 	failingAssert.Range(map[int]int{3: 1, 2: 2, 1: 3}, 5, 10, "map length out of range")
@@ -164,7 +169,7 @@ func TestOffsetPrint(t *testing.T) {
 	assert := audit.NewTestingAssertion(t, false)
 
 	// Log should reference line below (167).
-	failWithOffset(assert, "167")
+	failWithOffset(assert, "172")
 }
 
 // TestAssertSubstring tests the Substring() assertion.